refactor(): * moved all generic utilities under a 'utils' folder * extracted io (export) utils, actions & reducers (wip) * lots of related cleanups
