+++ src/io/reducers.js
@@ -1,0 +1,17 @@
+const {exportFilePathFromFormatAndDesign} = require('./exportUtils')
+
+const initialize = () => {
+  return {
+    exportFormat: '',
+    exportFilePath: '', // default export file path
+    availableExportFormats: []
+  }
+}
+const changeExportFormat = (state, exportFormat) => {
+  return Object.assign({}, state, exportFilePathFromFormatAndDesign(state.design, exportFormat))
+}
+
+module.exports = {
+  initialize,
+  changeExportFormat
+}
